Interactive mode support
It allows to execute Rubeus in an interactive mode. It is useful to avoid the command line process auditing functionality.

It switches to interactive mode only when no command line arguments have been input as mimikatz does.

It is also useful if you want to use it with loaders which inject the process into memory such as PEzor. Although Donut already allows hardcoding arguments, it is not efficient generating a binary for each command.

This condition (&& args[0] != "") is just to make it work properly with PEzor. I don't know why but the loader adds an empty first argument. It has no impact on the normal behavior.